[Refactor] Change folder and file structure for adding store folder
