+++ src/Logout/LogoutPage.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { useSelector, useDispatch } from 'react-redux';
 
-import { logout } from '../actions';
+import { logout } from '../store/actions';
 
 import { get } from '../utils';
